+++ src/llamafactory/v1/plugins/model_plugins/kernels/mlp/npu_swiglu.py
@@ -11,14 +11,15 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import importlib.util
+
 import re
 import types
 
 import torch
 
-from .....extras.types import DeviceType, HFModel, KernelType
+from .....extras.types import HFModel
 from ....trainer_plugins.distributed.accelerate import is_torch_npu_available
+from ..constants import DeviceType, KernelType
 from ..registry import KERNEL_REGISTRY, MetaSwiGluKernel
 
 
@@ -43,7 +44,7 @@ class NpuSwiGluKernel(MetaSwiGluKernel):
 
     @classmethod
     def apply(cls, model, **kwargs) -> 'HFModel':
-        if not (is_torch_npu_available() and importlib.util.find_spec("torch_npu")):
+        if not is_torch_npu_available():
             return model
 
         swiglu_pattern = re.compile("MLP", re.IGNORECASE)
