+++ src/llamafactory/v1/plugins/model_plugins/kernels/registry.py
@@ -11,11 +11,13 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
 from abc import ABC, abstractmethod
 from typing import Any, Callable, Optional
 
-from ....extras.types import DeviceType, HFModel, KernelType
+from ....extras.types import HFModel
 from ...trainer_plugins.distributed.accelerate import get_available_accelerator
+from .constants import DeviceType, KernelType
 
 
 class KernelRegistry:
