+++ src/llamafactory/v1/plugins/model_plugins/kernels/rms_norm/npu_rms_norm.py
@@ -11,12 +11,12 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import importlib.util
 import re
 import types
 
-from .....extras.types import DeviceType, HFModel, KernelType
+from .....extras.types import HFModel
 from ....trainer_plugins.distributed.accelerate import is_torch_npu_available
+from ..constants import DeviceType, KernelType
 from ..registry import KERNEL_REGISTRY, MetaRMSNormKernel
 
 
@@ -57,7 +57,7 @@ class NpuRMSNormKernel(MetaRMSNormKernel):
         - Do not modify weights, hyperparameters, or module structure to ensure
           numerical behavior and interface consistency.
         """
-        if not (is_torch_npu_available() and importlib.util.find_spec("torch_npu")):
+        if not is_torch_npu_available():
             return model
 
         rms_norm_pattern = re.compile("RMSNorm", re.IGNORECASE)
